updated README.md
I am going to upload to the github repo README.md file what I have written so far about our project. Feel free to read it over and use pieces of it in the Jupyter Notebook. I haven't finished it yet I'm still working on the LOAD section where we describe "the final database, tables/collections, and why this was chosen." I believe we are using the relational database pgAdmin4 for this?